+++ R/f7Skeleton.R
@@ -2,9 +2,15 @@
 #'
 #' Nice loading overlay for UI elements.
 #'
-#' @param tag Tag to be modified.
+#' @param target CSS selector on which to apply the effect.
+#' In general, you apply the effect on a wrapper such as a card,
+#' such that all nested elements receive the skeleton.
 #' @param effect Choose between "fade", "blink" or "pulse".
-#' @param duration Effect duration: 2s by default.
+#' @param duration Effect duration. NULL by default. If you know
+#' exactly how much time your most time consuming output takes
+#' to render, you can pass an explicit duration. In other cases,
+#' leave it to NULL.
+#' @param session Shiny session object.
 #'
 #' @export
 #'
@@ -27,48 +33,42 @@
 #'            f7Button(color = "blue", label = "My button", href = "https://www.google.com"),
 #'            f7Badge("Badge", color = "green")
 #'          )
-#'        ) %>% f7Skeleton(),
+#'        ),
 #'
 #'         f7List(
 #'          f7ListItem(
 #'            href = "https://www.google.com",
 #'            title = "Item 1"
-#'          ) %>% f7Skeleton(effect = "pulse", duration = 5) ,
+#'          ),
 #'          f7ListItem(
 #'            href = "https://www.google.com",
 #'            title = "Item 2"
-#'          ) %>% f7Skeleton(effect = "pulse", duration = 5)
+#'          )
 #'         )
 #'      )
 #'    ),
-#'    server = function(input, output) {}
+#'    server = function(input, output, session) {
+#'      observeEvent(TRUE, {
+#'        f7Skeleton(".card", "fade", 2)
+#'      }, once = TRUE)
+#'    }
 #'  )
 #' }
-f7Skeleton <- function(tag, effect = "fade", duration = 2) {
+f7Skeleton <- function(
+    target = ".card",
+    effect = c("fade", "blink", "pulse"),
+    duration = NULL,
+    session = shiny::getDefaultReactiveDomain()
+) {
 
-  id <- paste0("skeleton_", round(stats::runif(1, min = 0, max = 1e9)))
-  # need to add an id to the selected tag to make sure
-  # that the effect is not applied to other tags
-  tag$attribs$id <- id
-
-  shiny::tagList(
-    shiny::singleton(
-      shiny::tags$script(
-        paste0(
-          "$(function() {
-            // add skeleton class
-            $('#", id, "').children().addClass('skeleton-text skeleton-effect-", effect, "');
-            $('#", id, " .badge, #", id, " .button').addClass('skeleton-block');
-            // after several seconds, remove it
-            setTimeout(function() {
-              $('#", id, "').children().removeClass('skeleton-text skeleton-effect-", effect, "');
-              $('#", id, " .badge, #", id, " .button').removeClass('skeleton-block');
-            }, ", duration * 1000, ");
-        });
-        "
-        )
-      )
-    ),
-    tag
+  effect <- match.arg(effect)
+  message <-dropNulls(
+    list(
+      target = target,
+      effect = effect,
+      duration = duration
+    )
   )
+
+  sendCustomMessage("add_skeleton", message, session)
 }